refactor: package 추가 및 LottoNumber class getter method제거
package 추가(io.mwkwon.lotto)
불필요한 파일 제거(empty.txt)
LottoNumber 리펙토링에 따른 test 코드 수정
