+++ README.md
@@ -43,7 +43,7 @@
 </pre>
 
 ## 구현 기능 목록
-* [ ] 로또 번호 자동 생성 기능
+* [x] 로또 번호 자동 생성 기능
     * 1에서 45까지 중복되지않는 6개의 숫자 생성 기능
     * 1에서 45까지 범위안의 숫자인지 유효성 검사하는 기능
 * [ ] 입력한 구입금액 기준 1장 이상의 로또 구매 기능
