--- src/main/java/LottoNumber.java
@@ -1,22 +1,0 @@
-public class LottoNumber {
-
-    public static final int MIN_LOTTO_NUMBER = 1;
-    public static final int MAX_LOTTO_NUMBER = 45;
-    public static final String EXCEPTION_MESSAGE = "1에서 45사이의 값만 입력 가능합니다.";
-    private int number;
-
-    public LottoNumber(int number) {
-        validateLottoNumberBound(number);
-        this.number = number;
-    }
-
-    private void validateLottoNumberBound(int number) {
-        if (number < MIN_LOTTO_NUMBER || number > MAX_LOTTO_NUMBER) {
-            throw new IllegalArgumentException(EXCEPTION_MESSAGE);
-        }
-    }
-
-    public int getNumber() {
-        return number;
-    }
-}
