+++ src/main/java/io/mwkwon/lotto/domain/LottoNumber.java
@@ -1,0 +1,36 @@
+package io.mwkwon.lotto.domain;
+
+import java.util.Objects;
+
+public final class LottoNumber {
+
+    public static final int MIN_LOTTO_NUMBER = 1;
+    public static final int MAX_LOTTO_NUMBER = 45;
+    private static final String EXCEPTION_MESSAGE = "1에서 45사이의 값만 입력 가능합니다.";
+
+    private final int number;
+
+    public LottoNumber(int number) {
+        validateLottoNumberBound(number);
+        this.number = number;
+    }
+
+    private void validateLottoNumberBound(final int number) {
+        if (number < MIN_LOTTO_NUMBER || number > MAX_LOTTO_NUMBER) {
+            throw new IllegalArgumentException(EXCEPTION_MESSAGE);
+        }
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        LottoNumber that = (LottoNumber) o;
+        return number == that.number;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(number);
+    }
+}
